ci: Generalize OS setup in one action
Use just one action for OS setup and make it call the right script. This
makes it easier to maintain as when devs are working with containers,
they can just call the script, and it also avoids repeated conditional
steps in the workflows.

If a distro is not supported, the action will simply fail and support
may be added later.

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>